+++ app/src/mobile/java/com/ssmhdssmhd/mxboxs/ui/fragment/SettingFragment.java
@@ -105,6 +105,7 @@ public class SettingFragment extends BaseFragment implements ConfigListener, Sit
         mBinding.themeColorText.setText(getThemeText());
         mBinding.dohText.setText(getDohList()[getDohIndex()]);
         mBinding.incognitoText.setText(Setting.getSwitch(Setting.isIncognito()));
+        mBinding.wallSoundText.setText(Setting.getSwitch(Setting.getWallSound()));
         mBinding.sizeText.setText((size = ResUtil.getStringArray(R.array.select_size))[PlayerSetting.getSize()]);
         String prefix = Setting.getParseServerPrefix();
         mBinding.parseServerText.setText(prefix.isEmpty() ? getString(R.string.setting_off) : prefix);
@@ -147,6 +148,7 @@ public class SettingFragment extends BaseFragment implements ConfigListener, Sit
         mBinding.wallDefault.setOnClickListener(this::setWallDefault);
         mBinding.wallRefresh.setOnClickListener(this::setWallRefresh);
         mBinding.wallRefresh.setOnLongClickListener(this::onWallHistory);
+        mBinding.wallSound.setOnClickListener(this::setWallSound);
     }
 
     @Override
@@ -292,6 +294,12 @@ public class SettingFragment extends BaseFragment implements ConfigListener, Sit
     private void setIncognito(View view) {
         Setting.putIncognito(!Setting.isIncognito());
         mBinding.incognitoText.setText(Setting.getSwitch(Setting.isIncognito()));
+    }
+
+    private void setWallSound(View view) {
+        Setting.putWallSound(!Setting.getWallSound());
+        mBinding.wallSoundText.setText(Setting.getSwitch(Setting.getWallSound()));
+        ConfigEvent.common();
     }
 
     private void setSize(View view) {